backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
anthropic>=0.49.0
openai>=1.0.0
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
pypdf>=5.0.0
python-pptx>=1.0.0
python-multipart>=0.0.16
httpx>=0.28.0
reportlab>=4.2.0
